+++ autotrade/src/main/java/br/com/autotrade/services/SalesRecordsServices.java
@@ -1,12 +1,14 @@
 package br.com.autotrade.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.autotrade.dtos.SalesRecordDTO;
+import br.com.autotrade.models.Buyer;
 import br.com.autotrade.models.SalesRecords;
 import br.com.autotrade.repository.SalesRecordsRepository;
 import jakarta.validation.Valid;
@@ -36,7 +38,7 @@ public class SalesRecordsServices {
 			SalesRecords createdSalesRecord = salesRecordsRepository.save(salesRedord);
 			return createdSalesRecord.getId();
 		} catch (Exception e) {
-			throw new Exception("Erro ao inserir novo Cliente");
+			throw new Exception("Erro ao inserir nova Venda");
 		}
 
 	}
@@ -44,7 +46,7 @@ public class SalesRecordsServices {
 	public Long register(SalesRecordDTO salesRecordDTO) throws Exception {
 		try {
 			if (salesRecordsRepository.findById(salesRecordDTO.getId()) != null) {
-				throw new Exception("Cliente já cadastrada");
+				throw new Exception("Venda já cadastrada");
 			}
 			System.out.println(salesRecordDTO);
 			SalesRecords salesRedords = mapper.map(salesRecordDTO, SalesRecords.class);
@@ -58,6 +60,15 @@ public class SalesRecordsServices {
 	}
 
 	public void deleteSalesRecords(Long id) throws Exception {
+		Optional<SalesRecords> salesRecords;
+		try {
+			salesRecords = salesRecordsRepository.findById(id);
+		} catch (Exception e) {
+			throw new Exception("Erro ao excluir");
+		}
+		if (salesRecords.isEmpty()) {
+			throw new Exception("Registro de venda não existente");
+		}
 
 		try {
 			salesRecordsRepository.deleteById(id);
@@ -66,15 +77,24 @@ public class SalesRecordsServices {
 		}
 	}
 
-	public SalesRecords searchAnSalesRecordById(Long id) {
-		return mapper.map(salesRecordsRepository.findById(id), SalesRecords.class);
+	public SalesRecords searchAnSalesRecordById(Long id) throws Exception {
+		Optional<SalesRecords> salesRecords = salesRecordsRepository.findById(id);
+		if(salesRecords.isEmpty()) {
+			throw new Exception("Não ha registro de venda com esse id");
+		}
+		return mapper.map(salesRecords, SalesRecords.class);
 	}
 
-	public SalesRecordDTO searchAnSalesRecordDTOById(Long id) {
-		return mapper.map(salesRecordsRepository.findById(id), SalesRecordDTO.class);
+	public SalesRecordDTO searchAnSalesRecordDTOById(Long id) throws Exception {
+		
+		Optional<SalesRecords> salesRecords = salesRecordsRepository.findById(id);
+		if(salesRecords.isEmpty()) {
+			throw new Exception("Não ha registro de venda com esse id");
+		}
+		return mapper.map(salesRecords, SalesRecordDTO.class);
 	}
 
-	public Long updateEmployee(SalesRecordDTO salesRecordDTO) throws Exception {
+	public Long updateSalesRecord(SalesRecordDTO salesRecordDTO) throws Exception {
 		SalesRecords salesRecords = mapper.map(salesRecordDTO, SalesRecords.class);
 		return save(salesRecords);
 	}
